// src/components/exhibition/dto/conference.dto.ts
/**
 * COMPAON API
 * API definition for COMPAON
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: compaon@dalavina.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ConferenceVideo } from './conference-video.dto';
import { ConferenceImage } from './conference-image.dto';

export interface Conference {
    readonly id: number;
    name: string;
    conference_template_id?: number;
    exhibition_id?: number;
    conference_images?: Array<ConferenceImage>;
    conference_videos?: Array<ConferenceVideo>;
}